requirements.txt
aiofiles==0.7.0
autoflake==1.4
black==21.8b0
dbus-python==1.2.16
fastapi==0.68.1
httpx==0.19.0
isort==5.9.3
Jinja2==3.0.1
mypy==0.910
psutil==5.8.0
gunicorn==20.1.0
uvicorn==0.15.0
pytest==6.2.5
python-dotenv==0.19.0
flake8==3.9.2